Add files via upload
I define a search node of the game to be a board, the number of moves made to reach the board, and the previous search node. First, an initial search node (the initial board, 0 moves, and a null previous search node) is inserted into a priority queue. Then, the search node with the minimum priority is deleted from the priority queue, and all neighboring search nodes (those that can be reached in one move from the dequeued search node) are inserted onto the priority queue . This procedure is repeated until the search node dequeued corresponds to a goal board. The success of this approach hinges on the choice of priority function for a search node. Two priority functions are considered:

Hamming priority function: The number of tiles in the wrong position, plus the number of moves made so far to get to the search node. Intuitively, a search node with a small number of tiles in the wrong position is close to the goal, and a search node that has been reached using a small number of moves i